mpu6000 split into separate main, header, implementation
