+++ src/drivers/imu/mpu6000/MPU6000_SPI.cpp
@@ -46,7 +46,7 @@
 #include <drivers/drv_accel.h>
 #include <drivers/drv_device.h>
 
-#include "mpu6000.h"
+#include "MPU6000.hpp"
 
 #include <string.h>
